Moves the setup of CEED Operator out of RHSFunction
The CEED Operator is now setup at the end of RDySetup instead of
setting it up when RHSFunction is called for the first time.
