+++ src/rdysetup.c
@@ -2,6 +2,7 @@
 #include <petscdmplex.h>
 #include <private/rdycoreimpl.h>
 #include <private/rdymemoryimpl.h>
+#include "swe_setup_ceed.h"
 #include <rdycore.h>
 
 // Maximum length of the name of a prognostic or diagnostic field component
@@ -1022,6 +1023,11 @@ PetscErrorCode RDySetup(RDy rdy) {
 
   RDyLogDebug(rdy, "Initializing solution data...");
   PetscCall(InitSolution(rdy));
+
+  if (rdy->ceed_resource[0]) {
+    RDyLogDebug(rdy, "Setting up the CEED Operator...");
+    PetscCall(RDyCeedOperatorSetUp(rdy));
+  }
 
   PetscFunctionReturn(0);
 }
